Starting to move mock clients into their own file
