+++ src/tests/index.test.ts
@@ -1,21 +1,32 @@
 import { it, describe, expect, beforeEach, afterEach, vi } from "vitest";
 import { SecurityHubJiraSync } from "../index";
-import { IAMClient, ListAccountAliasesCommand } from "@aws-sdk/client-iam";
-import {
-  SecurityHubClient,
-  GetFindingsCommand,
-} from "@aws-sdk/client-securityhub";
-import { STSClient, GetCallerIdentityCommand } from "@aws-sdk/client-sts";
-import { mockClient } from "aws-sdk-client-mock";
 import JiraClient, { IssueObject, JsonResponse } from "jira-client";
 import { Jira } from "../libs";
 import axios, { AxiosRequestConfig } from "axios";
+import { mockClients } from "./mockClients";
 import * as constants from "./constants";
 import * as mockResponses from "./mockResponses";
 
 // ******** mocks ********
-const jiraAddNewIssueCalls: IssueObject[] = [];
-const jiraSearchCalls: JsonResponse[] = [];
+let jiraAddNewIssueCalls: IssueObject[] = [];
+let jiraSearchCalls: JsonResponse[] = [];
+let originalJiraClosedStatuses;
+
+beforeEach(() => {
+  process.env.JIRA_HOST = "testHost";
+  process.env.JIRA_USERNAME = "testUsername";
+  process.env.JIRA_TOKEN = "testToken";
+  process.env.JIRA_PROJECT = constants.testProject;
+  process.env.JIRA_CLOSED_STATUSES = constants.testStatus;
+  mockClients();
+  jiraAddNewIssueCalls = [];
+  jiraSearchCalls = [];
+  originalJiraClosedStatuses = process.env.JIRA_CLOSED_STATUSES;
+});
+
+afterEach(() => {
+  process.env.JIRA_CLOSED_STATUSES = originalJiraClosedStatuses;
+});
 
 vi.mock("jira-client", () => {
   return {
@@ -48,64 +59,6 @@ vi.mock("axios", () => {
       return axiosInstance.request(config);
     },
   };
-});
-
-// IAM
-const iamClient = mockClient(IAMClient);
-iamClient
-  .on(ListAccountAliasesCommand, {})
-  .resolves(mockResponses.listAccountAliasesResponse);
-
-// Security Hub
-const sHClient = mockClient(SecurityHubClient);
-sHClient
-  .on(GetFindingsCommand, {})
-  .resolvesOnce({
-    ...mockResponses.getFindingsCommandResponse,
-    NextToken: "test",
-  })
-  .resolves({
-    ...mockResponses.getFindingsCommandResponse,
-    ...{
-      Findings: [
-        {
-          ...mockResponses.getFindingsCommandResponse.Findings[0],
-          ProductFields: {
-            Title: "Test Finding",
-            StandardsControlArn: `arn:aws:securityhub:${constants.testAwsRegion}:${constants.testAwsAccountId}:control/aws-foundational-security-best-practices/v/1.0.0/KMS.3`,
-          },
-        },
-      ],
-    },
-  });
-
-// STS
-const stsClient = mockClient(STSClient);
-stsClient.on(GetCallerIdentityCommand, {}).resolves({
-  Account: constants.testAwsAccountId,
-});
-
-// ******** setup ********
-process.env.JIRA_HOST = "testHost";
-process.env.JIRA_USERNAME = "testUsername";
-process.env.JIRA_TOKEN = "testToken";
-process.env.JIRA_PROJECT = constants.testProject;
-process.env.JIRA_CLOSED_STATUSES = constants.testStatus;
-
-let originalJiraClosedStatuses;
-
-beforeEach(() => {
-  originalJiraClosedStatuses = process.env.JIRA_CLOSED_STATUSES;
-  iamClient.resetHistory();
-  stsClient.resetHistory();
-
-  // Reset the calls arrays
-  jiraSearchCalls.length = 0;
-  jiraAddNewIssueCalls.length = 0;
-});
-
-afterEach(() => {
-  process.env.JIRA_CLOSED_STATUSES = originalJiraClosedStatuses;
 });
 
 // ******** tests ********
